various changes
Biggest changes in the main were to move the character movement to a separate class.  The event handler is still in the main. The next biggest change is having dialog boxes pop up in both the left and rick click interaction methods for ores. Also changed all my variable names in main since they sucked